+++ src/main/kotlin/local/mathewdj/tennis/repository/SinglesMatchRepository.kt
@@ -3,4 +3,4 @@ package local.mathewdj.tennis.repository
 import local.mathewdj.tennis.entity.SinglesMatchEntity
 import org.springframework.data.repository.CrudRepository
 
-interface SinglesMatchRepository : CrudRepository<SinglesMatchEntity, Long>
+interface SinglesMatchRepository : CrudRepository<SinglesMatchEntity, Int>
